implementa classe controller Conta, metodo post conta

+++ src/main/java/br/com/hoythusheim/dto/ContaDto.java
@@ -2,6 +2,7 @@ package br.com.hoythusheim.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Builder
@@ -9,7 +10,7 @@ public class ContaDto {
 
     private Long id;
     private String emissor;
-    private Long valor;
+    private BigDecimal valor;
     private Date diaRecebimento;
     private Date diaPagamento;
     private String statusConta;
